requirements.txt
numpy==1.21.0
norfair==0.3.1
torch==1.9.0
pandas==1.3.2
requests==2.26.0
pyyaml==5.4.1
tqdm==4.62.0
torchvision==0.10.0
seaborn==0.11.0
matplotlib==3.4.2
imutils==0.5.4
python-dateutil==2.8.2
tinytag==1.5.0